Initial push
Alexandre Brisebois' Backup And Restore Tool for Windows Azure Blob
Storage Service

Disclaimer

These are my personal Windows Azure tools. I am actively adding to them
as I publish new blog posts. This code does not come with any
guaranties. Bugs are possible and I will be happy to fix bugs that are
reported through the dedicated Issues page on the GitHub repository.
Please keep in mind that work on this code is done on me personal time.
For those who are wish to contribute bug fixes and new features, I will
be happy to consider all pull requests and collaborate in order to
integrate your generous contributions to the package